+++ DailyIncrementalBackup.ps1
@@ -1,4 +1,5 @@
 ﻿#requires -version 5.1
+#requires -module BurntToast
 
 #backup files from CSV change logs
 
@@ -9,7 +10,7 @@ foreach ($name in (Import-CSV D:\Backup\Scripts-log.csv -OutVariable in | Select
  $in | where {$_.name -EQ $name} | sort-object -Property ID | Select-Object -last 1
 }
 #>
-Write-Host "[(Get-Date)] Starting $($myinvocation.MyCommand)]" -fore Cyan
+Write-Host "[$(Get-Date)] Starting $($myinvocation.MyCommand)]" -fore Cyan
 
 #this is my internal archiving code. You can use whatever you want.
 Import-Module C:\scripts\PSRAR\Dev-PSRar.psm1 -force
@@ -19,8 +20,9 @@ $paths = (Get-ChildItem -Path D:\Backup\*.csv).Fullname
 foreach ($path in $paths) {
 
     $name = (Split-Path -Path $Path -Leaf).split("-")[0]
-    $files = Import-Csv -path $path | Where-Object { ($_.name -notmatch "~|\.tmp") -AND ($_.size -gt 0) -AND ($_.IsFolder -eq 'False') -AND (Test-Path $_.path) } |
-        Select-Object -Property path, size, directory, isfolder, ID | Group-Object -Property Path
+    $files = Import-Csv -path $path |
+    Where-Object { ($_.name -notmatch "~|\.tmp") -AND ($_.size -gt 0) -AND ($_.IsFolder -eq 'False') -AND (Test-Path $_.path) } |
+    Select-Object -Property path, size, directory, isfolder, ID | Group-Object -Property Path
 
     foreach ($file in $files) {
         $parentFolder = $file.group[0].directory
@@ -28,14 +30,14 @@ foreach ($path in $paths) {
         $relPath = Join-Path -Path "D:\backtemp\$Name" -childpath $parentFolder.Substring(3)
 
         if (-Not (Test-Path -path $relpath)) {
-            Write-Host "[(Get-Date)] Creating $relpath" -ForegroundColor cyan
+            Write-Host "[$(Get-Date)] Creating $relpath" -ForegroundColor cyan
             $new = New-Item -path $relpath -ItemType directory -force
             Start-Sleep -Milliseconds 100
 
             #copy hidden attributes
             $attrib = (Get-Item $parentfolder -force).Attributes
             if ($attrib -match "hidden") {
-                Write-Host "[(Get-Date)] Copying attributes from $parentfolder to $($new.FullName)" -ForegroundColor yellow
+                Write-Host "[$(Get-Date)] Copying attributes from $parentfolder to $($new.FullName)" -ForegroundColor yellow
                 Write-Host $attrib -ForegroundColor yellow
                 (Get-Item $new.FullName -force).Attributes = $attrib
             }
@@ -49,15 +51,31 @@ foreach ($path in $paths) {
     #create a RAR archive or substitute your archiving code
     $archive = Join-Path D:\BackTemp -ChildPath "$(Get-Date -Format yyyyMMdd)_$name-INCREMENTAL.rar"
     Add-RARContent -Object $relPath -Archive $archive -CompressionLevel 5 -Comment "Incremental backup $(Get-Date)"
-    Write-Host "[(Get-Date)] Moving $archive to NAS" -fore green
+    Write-Host "[$(Get-Date)] Moving $archive to NAS" -fore green
     Move-Item -Path $archive -Destination \\ds416\backup -Force
 
-    Write-Host "[(Get-Date)] Removing $path" -fore yellow
+    Write-Host "[$(Get-Date)] Removing $path" -fore yellow
     Remove-Item $path
 
 } #foreach path
 
-Write-Host "[(Get-Date)] Removing temporary Backup folders" -fore yellow
+Write-Host "[$(Get-Date)] Removing temporary Backup folders" -fore yellow
 Get-Childitem -path D:\BackTemp -directory | Remove-Item -Force -Recurse
 
-Write-Host "[(Get-Date)] Ending $($myinvocation.MyCommand)]" -for cyan
+$newFiles = Get-Childitem -Path \\ds416\backup\*incremental.rar | Where-Object LastWriteTime -ge (Get-Date).Date
+#send a toast notification
+$btText = @"
+Backup Task Complete
+
+"@
+$newfiles | Foreach-Object { $btText+= "$($_.name)`n"}
+
+$params = @{
+    Text    = $btText
+    Header  = $(New-BTHeader -Id 1 -Title "Daily Incremental Backup")
+    Applogo = "c:\scripts\db.png"
+}
+
+New-BurntToastNotification @params
+
+Write-Host "[$(Get-Date)] Ending $($myinvocation.MyCommand)]" -for cyan